App Improve

+++ src/schemas/ahorro.schema.ts
@@ -5,7 +5,7 @@ const fechaRegex = /^\d{4}-\d{2}-\d{2}$/;
 export const ahorroBaseSchema = z.object({
   periodo_id: z.number().int().positive(),
   descripcion: z.string().trim().min(1).max(255),
-  monto: z.number().positive(),
+  monto: z.number().int().positive(),
   moneda: z.enum(['ARS', 'USD']),
   origen: z.string().trim().max(255).nullable().optional(),
   fecha: z.string().regex(fechaRegex, 'La fecha debe tener formato YYYY-MM-DD'),
